backend/requirements.txt
Django==5.0.3
djangorestframework==3.15.1
google-generativeai==0.7.2
python-dotenv==1.0.1
django-cors-headers==4.3.1
